Added data_prep - modules to prepare data from MIDI Format 0 to required data format
Generator and Discriminator - number of neurons changed again, LSTM are replaced to GRU

Tester - now we define not only first chord of sequence, but every X-th chord. It's like "markers" on the way of composition that helps network not to stuck on some chord. Fixed some bugs

Trainer: 1. Gloss function changed a bit. 2. Added bias to generator to make him to produce chords with less number of notes (not very successfull yet). 3. Still trying to make gen loss to fall down (also not very successfull).